+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## 0.2.0 (7 Nov 2018)
++ Adds a green spinner to accounts list and projects list
++ List projects in table form
+
 ## 0.1.0 (4 Nov 2018)
 + Skeleton Cobra commands without any pudding